Backend/JobMatcher/requirements.txt
# -----------------------------
# API Framework
# -----------------------------
fastapi==0.111.1
uvicorn[standard]==0.23.1

# -----------------------------
# LangChain ecosystem
# -----------------------------
langchain==1.0.4
langchain-community==0.4.1
langchain-huggingface==1.0.1

# -----------------------------
# AI / Embeddings
# -----------------------------
sentence-transformers==5.1.2
transformers==4.57.1

# -----------------------------
# Vector Database
# -----------------------------
chromadb==1.3.4

# -----------------------------
# PyTorch CPU-only (small build)
# -----------------------------
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.8.0+cpu
torchvision==0.23.0+cpu
torchaudio==2.8.0+cpu

# -----------------------------
# Google Cloud
# -----------------------------
google-cloud-firestore>=2.11.0,<3.0.0
google-auth>=2.15.0
